CDPCP-3702 Add flag to force update of workload credentials by user sync
Adds a flag to the SynchronizeAllUsers request that forces user sync to update
workload credentials for all users, whether or not they have changed. The flag
is a no-op when the CDP_USER_SYNC_CREDENTIALS_UPDATE_OPTIMIZATION entitlement
is not enabled.

The flag will be exposed as a '--force-workload-credentials-update' option to
the 'cdp environments sync-all-users' API. This is the backend implementation.

I added the flag to both public entry points to UserSyncService -- synchronizeUsers
and synchronizeUsersWithCustomPermissionCheck, even though callers of the former
always set it to false. It seemed inconsistent for this option to be accepted
when doing a custom permission check, but not otherwise.

I also moved the computation of UserSyncOptions up the call chain, to avoid
adding more parameters to lower-level methods, and to simplify adding new
options in the future.